% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.r
\name{add_deletion}
\alias{add_deletion}
\title{add_deletion}
\usage{
add_deletion(x, tree, mutation_site, config)
}
\arguments{
\item{x}{a character matrix}

\item{tree}{a matrix representing the lineage tree}

\item{mutation_site}{a binary matrix for mutation site}

\item{config}{a lineage_tree_config object}
}
\value{
a character matrix with deletions
}
\description{
Add deletion
}
