% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.r
\name{add_dropout}
\alias{add_dropout}
\title{add_dropout}
\usage{
add_dropout(x, config)
}
\arguments{
\item{x}{a character matrix}

\item{config}{a lineage_tree_config object}
}
\value{
a character matrix with dropout events
}
\description{
Add dropout events
}
