% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_igraph.r
\name{as_igraph,phylo-method}
\alias{as_igraph,phylo-method}
\title{as_igraph}
\usage{
\S4method{as_igraph}{phylo}(x)
}
\arguments{
\item{x}{a phylo object}
}
\value{
an igraph object
}
\description{
Convert an phylo object to an igraph object, while keeping the weight (in contrast to igraph::as.igraph)
}
