% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.r
\name{as_igraph}
\alias{as_igraph}
\title{Generic function for as_igraph}
\usage{
as_igraph(x, ...)
}
\arguments{
\item{x}{a phylo object}

\item{...}{additional parameters}
}
\description{
Generic function for as_igraph
}
