% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_lineage_tree.r
\name{as_lineage_tree,phyDat,phylo,lineage_tree_config-method}
\alias{as_lineage_tree,phyDat,phylo,lineage_tree_config-method}
\title{as_lineage_tree}
\usage{
\S4method{as_lineage_tree}{phyDat,phylo,lineage_tree_config}(x, y, config, ...)
}
\arguments{
\item{x}{a phyDat object}

\item{y}{a phylo object}

\item{config}{a lineage_tree_config object}

\item{...}{additional parameters}
}
\value{
a lineage_tree object
}
\description{
Convert a phylo object and a phyDat object to a lineage_tree object
}
