% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.r
\name{as_lineage_tree}
\alias{as_lineage_tree}
\title{Generic function for as_lineage_tree}
\usage{
as_lineage_tree(x, y, config, ...)
}
\arguments{
\item{x}{a phyDat object}

\item{y}{a phylo object}

\item{config}{a lineage_tree_config object}

\item{...}{additional parameters}
}
\description{
Generic function for as_lineage_tree
}
