% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_phylo.r
\name{as_phylo,igraph-method}
\alias{as_phylo,igraph-method}
\title{as_phylo}
\usage{
\S4method{as_phylo}{igraph}(x)
}
\arguments{
\item{x}{an igraph object}
}
\value{
a phylo object
}
\description{
Convert an igraph object to a phylo object
}
