% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_replacment_matrix.r
\name{compute_replacement_matrix,phyDat-method}
\alias{compute_replacement_matrix,phyDat-method}
\title{compute_replacement_matrix}
\usage{
\S4method{compute_replacement_matrix}{phyDat}(
  x,
  n_batch = 10L,
  n_cells_per_batch = 200L,
  k = 2,
  division = 20L,
  deletion = TRUE
)
}
\arguments{
\item{x}{a phyDat object}

\item{n_batch}{number of simulated trees}

\item{n_cells_per_batch}{number of sampled tip or internval nodes in the simulated tree}

\item{k}{k-mer}

\item{division}{number of cell division}

\item{deletion}{whether or not include the deletion events}
}
\value{
a data frame
}
\description{
Simulate replacement matrix with input sequence data
}
\author{
Wuming Gong (gongx030@umn.edu)
}
