% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_replacment_matrix.r
\name{compute_replacement_matrix_core}
\alias{compute_replacement_matrix_core}
\title{compute_replacement_matrix_core}
\usage{
compute_replacement_matrix_core(
  n_batch = 10L,
  n_cells_per_batch = 200L,
  k = 2,
  alphabets = NULL,
  outcome_prob = NULL,
  sequence_length = 200L,
  mutation_probs = seq(0.01, 0.3, by = 0.01),
  division = 20L,
  n_leaves = 200L,
  dropout = TRUE
)
}
\arguments{
\item{n_batch}{number of simulated trees}

\item{n_cells_per_batch}{number of sampled tip or internval nodes in the simulated tree}

\item{k}{k-mer}

\item{alphabets}{alphabets used in the tree}

\item{outcome_prob}{outcome probability of each letter}

\item{sequence_length}{sequence length (e.g. number of targets)}
}
\description{
compute_replacement_matrix_core
}
