% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCLEAR-package.r
\docType{package}
\name{DCLEAR}
\alias{DCLEAR}
\title{DCLEAR: A package for DCLEAR: Distance based Cell LinEAge Reconstruction}
\description{
Distance based methods for inferring lineage treess from single cell data
}
