% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_replacement.r
\name{dist_kmer_replacement_inference}
\alias{dist_kmer_replacement_inference}
\title{Core function of computing kmer replacement distance}
\usage{
dist_kmer_replacement_inference(x, kmer_summary, k = 2)
}
\arguments{
\item{x}{input data in phyDat format}

\item{kmer_summary}{a kmer_summary object}

\item{k}{k-mers (default k=2)}
}
\value{
a dist object
}
\description{
Compute the sequence distance matrix using inferred kmer replacement matrix
}
\author{
Wuming Gong (gongx030@umn.edu)
}
