% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_replacement.r
\name{dist_replacement,phyDat,dgCMatrix-method}
\alias{dist_replacement,phyDat,dgCMatrix-method}
\title{dist_replacement}
\usage{
\S4method{dist_replacement}{phyDat,dgCMatrix}(x, replacement_matrix, ...)
}
\description{
compute the kmer replacement distance matrix
}
