% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_replacement.r
\name{dist_replacement,phyDat,kmer_summary-method}
\alias{dist_replacement,phyDat,kmer_summary-method}
\title{dist_replacement}
\usage{
\S4method{dist_replacement}{phyDat,kmer_summary}(x, kmer_summary, k = 2, ...)
}
\arguments{
\item{kmer_summary}{a kmer_summary object}

\item{ix}{nput data in phyDat format}
}
\value{
a dist object
}
\description{
Compute the sequence distance by kmer replacement
}
\author{
Wuming Gong (gongx030@umn.edu)
}
