% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_replacement.r
\name{dist_replacement,phyDat,missing-method}
\alias{dist_replacement,phyDat,missing-method}
\title{Compute the kmer replacement distance}
\usage{
\S4method{dist_replacement}{phyDat,missing}(x, kmer_summary, k = 2, reps = 20L, division = 16L, ...)
}
\arguments{
\item{x}{input data in phyDat format}

\item{kmer_summary}{a kmer_summary object}
}
\value{
a dist object
}
\description{
Compute the kmer replacement distance between sequences
}
\author{
Wuming Gong (gongx030@umn.edu)
}
