% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.r
\name{dist_replacement}
\alias{dist_replacement}
\title{Generic function for dist_replacement}
\usage{
dist_replacement(x, kmer_summary, k, ...)
}
\arguments{
\item{x}{a sequence object}

\item{kmer_summary}{a kmer_summary object}

\item{k}{k-mer length}

\item{...}{additional parameters}
}
\description{
Generic function for dist_replacement
}
