% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_replacement.r
\name{dist_replacement_bayesian_core}
\alias{dist_replacement_bayesian_core}
\title{dist_replacement_bayesian_core}
\usage{
dist_replacement_bayesian_core(x, kmer_summary)
}
\description{
dist_replacement_bayesian_core
}
