% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downsample.r
\name{downsample,igraph-method}
\alias{downsample,igraph-method}
\title{downsample}
\usage{
\S4method{downsample}{igraph}(x, n = 10L, ...)
}
\arguments{
\item{x}{a igraph object}

\item{n}{number of leaves (tips) in the down-sampled tree}

\item{...}{additional parameters}
}
\value{
a phylo object
}
\description{
Sample a lineage tree
}
