% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downsample.r
\name{downsample,lineage_tree-method}
\alias{downsample,lineage_tree-method}
\title{downsample}
\usage{
\S4method{downsample}{lineage_tree}(x, n = 10L, ...)
}
\arguments{
\item{x}{a lineage_tree object}

\item{n}{number of leaves (tips) in the down-sampled tree}

\item{...}{additional parameters}
}
\value{
a lineage_tree object
}
\description{
Sample a lineage tree
}
