% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.r
\name{downsample}
\alias{downsample}
\title{Generic function for downsample}
\usage{
downsample(x, ...)
}
\arguments{
\item{x}{a data object}

\item{...}{additional parameters}
}
\description{
Generic function for downsample
}
