% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_replacement.r
\name{get_conditional_replacement_probability}
\alias{get_conditional_replacement_probability}
\title{get_conditional_replacement_probability}
\usage{
get_conditional_replacement_probability(x)
}
\description{
get_conditional_replacement_probability
}
