% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_replacement.r
\name{get_distance_prior}
\alias{get_distance_prior}
\title{get_distance_prior}
\usage{
get_distance_prior(x)
}
\arguments{
\item{x}{a kmer_summary object}
}
\value{
a probabilistic vector of the distribution of nodal distances
}
\description{
prior distribution of distance
}
\author{
Wuming Gong (gongx030@umn.edu)
}
