% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_replacement.r
\name{get_distance_probability}
\alias{get_distance_probability}
\title{get_distance_probability}
\usage{
get_distance_probability(x)
}
\description{
Compute the conditional distribution of distance according to k-mer replacement
}
