% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.r
\name{get_leaves}
\alias{get_leaves}
\title{Generic function for get_leaves}
\usage{
get_leaves(x, ...)
}
\arguments{
\item{x}{a lineage_tree object}

\item{...}{additional parameters}
}
\description{
Generic function for get_leaves
}
