% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.r
\name{get_node_names}
\alias{get_node_names}
\title{get_node_names}
\usage{
get_node_names(x)
}
\arguments{
\item{x}{node id}
}
\value{
node names
}
\description{
Convenient function for get node names
}
\author{
Wuming Gong (gongx030@umn.edu)
}
