% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_replacement.r
\name{get_replacement_probability}
\alias{get_replacement_probability}
\title{get_replacement_probability}
\usage{
get_replacement_probability(x)
}
\arguments{
\item{x}{a kmer_summary object}
}
\value{
an 3D probabilistic array (kmers by kmers by distances)
}
\description{
Compute p(A,B|d), the conditional probability of seeing a replacement of from kmer A to B or vice versa
}
\author{
Wuming Gong (gongx030@umn.edu)
}
