% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.r
\name{get_sequence}
\alias{get_sequence}
\title{get_sequence}
\usage{
get_sequence(x, tree, outcome, config)
}
\arguments{
\item{x}{a character matrix}

\item{tree}{a matrix representing the lineage tree}

\item{outcome}{a character matrix}

\item{config}{a lineage_tree_config object}
}
\value{
a character matrix
}
\description{
Get sequencees
}
