% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_replacement.r
\name{get_transition_probability}
\alias{get_transition_probability}
\title{get_transition_probability}
\usage{
get_transition_probability(x)
}
\arguments{
\item{x}{a kmer_summary object}
}
\value{
an 3D probabilistic array (kmers by kmers by distances)
}
\description{
Compute p(A,X|B,Y,d), the conditional probability of seeing a replacement from A to B
given the previous replacement B from Y at nodal distance d
}
\author{
Wuming Gong (gongx030@umn.edu)
}
