% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineages-data.r
\docType{data}
\name{lineages}
\alias{lineages}
\title{Lineage data}
\format{
An object of class \code{list} of length 100.
}
\usage{
data(lineages)
}
\description{
Lineage data
}
\examples{
data(lineages)
}
\keyword{datasets}
