% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.r
\name{positional_mutation_prob}
\alias{positional_mutation_prob}
\title{positional_mutation_prob}
\usage{
positional_mutation_prob(x, config)
}
\arguments{
\item{x}{a phyDat object}

\item{config}{a lineage_tree_config object}
}
\value{
a positional mutation probability matrix
}
\description{
Convenient function for get node names
}
