% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_sequence.r
\name{process_sequence,phyDat-method}
\alias{process_sequence,phyDat-method}
\title{Process sequences}
\usage{
\S4method{process_sequence}{phyDat}(x, division = 16L)
}
\arguments{
\item{x}{input data in phyDat format}

\item{division}{cell divisions (default: 16L)}
}
\value{
a `lineage_tree_config` object
}
\description{
Process sequences
}
\author{
Wuming Gong (gongx030@umn.edu)
}
