% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.r
\name{process_sequence}
\alias{process_sequence}
\title{Generic function for process_sequence}
\usage{
process_sequence(x, ...)
}
\arguments{
\item{x}{a sequence object}

\item{...}{additional parameters}
}
\value{
a `lineage_tree_config` object
}
\description{
Generic function for process_sequence
}
