% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prune.r
\name{prune,igraph-method}
\alias{prune,igraph-method}
\title{prune}
\usage{
\S4method{prune}{igraph}(x, weighted = TRUE, ...)
}
\arguments{
\item{x}{an igraph object}

\item{weighted}{whether or not keep the edge weight (default: TRUE)}

\item{...}{additional parameters}
}
\value{
an igraph object
}
\description{
Trim a full lineage tree into phylogenetic tree
}
