% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.r
\name{prune}
\alias{prune}
\title{Generic function for prune}
\usage{
prune(x, ...)
}
\arguments{
\item{x}{a lineage_tree object}

\item{...}{additional parameters}
}
\description{
Generic function for prune
}
