% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.r
\name{random_tree}
\alias{random_tree}
\title{random_tree}
\usage{
random_tree(n_samples, division = 16L)
}
\arguments{
\item{n_samples}{number of samples to simulate}

\item{division}{number of cell division}
}
\value{
a data frame
}
\description{
Simulate a random lineage tree
}
\author{
Wuming Gong (gongx030@umn.edu)
}
