% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.r
\name{rbind,phyDat-method}
\alias{rbind,phyDat-method}
\title{rbind}
\usage{
\S4method{rbind}{phyDat}(..., deparse.level = 1)
}
\arguments{
\item{...}{additional parameters}

\item{x}{a lineage_tree object}
}
\value{
a phyData object
}
\description{
Get the leaf sequences
}
