% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.r
\name{sample_mutation_outcome}
\alias{sample_mutation_outcome}
\title{sample_mutation_outcome}
\usage{
sample_mutation_outcome(x, mp = NULL, config)
}
\arguments{
\item{x}{an igraph object}

\item{mp}{a mutation site matrix}

\item{config}{a lineage_tree_config object}
}
\value{
a outcome matrix
}
\description{
Sample mutation outcome
}
