% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.r
\name{sample_mutation_site}
\alias{sample_mutation_site}
\title{sample_mutation_site}
\usage{
sample_mutation_site(tree, config)
}
\arguments{
\item{tree}{a data frame}

\item{config}{a lineage_tree_config object}
}
\value{
a mutation site matrix
}
\description{
Sample mutation site
}
