% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.r
\name{score_simulation}
\alias{score_simulation}
\title{score_simulation}
\usage{
score_simulation(x, y, config)
}
\arguments{
\item{x}{a character matrix}

\item{y}{a character matrix}

\item{config}{a lineage_tree_config object}
}
\value{
numeric scores
}
\description{
Compare two sets of sequences
}
