% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.r
\name{simulate,lineage_tree_config-method}
\alias{simulate,lineage_tree_config-method}
\title{simulate}
\usage{
\S4method{simulate}{lineage_tree_config}(config, n_samples = 200)
}
\arguments{
\item{config}{simulation configuration; a lineage_tree_config object}

\item{n_samples}{number of samples to simulate}
}
\value{
a lineage_tree object
}
\description{
Simulate a cell lineage tree
Adoped from https://github.com/elifesciences-publications/CRISPR_recorders_sims/blob/master/MATLAB_sims/GESTALT_30hr_1x_simulation.m
}
\author{
Wuming Gong (gongx030@umn.edu)
}
