% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.r
\name{simulate,lineage_tree_config,phyDat-method}
\alias{simulate,lineage_tree_config,phyDat-method}
\title{simulate}
\usage{
\S4method{simulate}{lineage_tree_config,phyDat}(config, x, n_samples = 200L, k = 50, greedy = TRUE, ...)
}
\arguments{
\item{config}{simulation configuration; a lineage_tree_config object}

\item{x}{a sequence object}

\item{n_samples}{number of samples to simulate}

\item{k}{Number of trials}

\item{greedy}{Whether ot not use a greedy search}

\item{...}{additional parameters}
}
\value{
a lineage_tree object
}
\description{
Simulate a cell lineage tree based on a set of sequences
}
\author{
Wuming Gong (gongx030@umn.edu)
}
