% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.r
\name{simulate}
\alias{simulate}
\title{simulate}
\usage{
simulate(
  n_samples = 200,
  n_targets = 200,
  mutation_prob = 0.03,
  division = 16L,
  alphabets = NULL,
  outcome_prob = NULL,
  deletion = TRUE
)
}
\arguments{
\item{n_samples}{number of samples to simulate}

\item{n_targets}{number of targets (e.g. sequence length)}

\item{mutation_prob}{mutation probability}

\item{division}{number of cell division}

\item{alphabets}{alphabets used in the tree}

\item{outcome_prob}{outcome probability of each letter}

\item{deletion}{whether or not include the deletion events}
}
\value{
a lineage_tree object
}
\description{
Simulate a cell lineage tree
Adoped from https://github.com/elifesciences-publications/CRISPR_recorders_sims/blob/master/MATLAB_sims/GESTALT_30hr_1x_simulation.m
}
\author{
Wuming Gong (gongx030@umn.edu)
}
