% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.r
\name{simulate_core}
\alias{simulate_core}
\title{simulate_core}
\usage{
simulate_core(config, mp = NULL, n_samples = 200L, ...)
}
\arguments{
\item{config}{simulation configuration; a lineage_tree_config object}

\item{mp}{site specific mutation probability}

\item{n_samples}{number of samples to simulate}

\item{...}{additional parameters}
}
\description{
Simulate a cell lineage tree
Adoped from https://github.com/elifesciences-publications/CRISPR_recorders_sims/blob/master/MATLAB_sims/GESTALT_30hr_1x_simulation.m
}
