% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.r
\name{simulate_targets}
\alias{simulate_targets}
\title{simulate_targets}
\usage{
simulate_targets(
  x,
  n_samples = 200,
  mutation_prob = 0.03,
  division = 16L,
  dropout = TRUE
)
}
\description{
simulate_targets
}
