% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/substr_kmer.r
\name{substr_kmer,kmer_summary-method}
\alias{substr_kmer,kmer_summary-method}
\title{Subseting a kmer_summary object}
\usage{
\S4method{substr_kmer}{kmer_summary}(x, k = 2)
}
\arguments{
\item{x}{a kmer_summary object}

\item{k}{k-mer length(default: 2)}
}
\value{
a new kmer_summary object
}
\description{
Summarize the short k-mer summary from the long k-mer summary
}
\author{
Wuming Gong (gongx030@umn.edu)
}
