% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.r
\name{substr_kmer}
\alias{substr_kmer}
\title{Generic function for substr_kmer}
\usage{
substr_kmer(x, ...)
}
\arguments{
\item{x}{a kmer object}

\item{...}{additional parameters}
}
\value{
a new kmer_summary object
}
\description{
Generic function for substr_kmer
}
