% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtract.r
\name{subtract,lineage_tree,lineage_tree-method}
\alias{subtract,lineage_tree,lineage_tree-method}
\title{subtract}
\usage{
\S4method{subtract}{lineage_tree,lineage_tree}(x, y, ...)
}
\arguments{
\item{x}{a lineage_tree object}

\item{y}{a lineage_tree object}

\item{...}{additional parameters}
}
\value{
a lineage_tree object
}
\description{
Subtract a subtree from a large tree
}
