% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.r
\name{subtract}
\alias{subtract}
\title{Generic function for subtract}
\usage{
subtract(x, y, ...)
}
\arguments{
\item{x}{a lineage_tree object}

\item{y}{a lineage_tree object}

\item{...}{additional parameters}
}
\description{
Generic function for subtract
}
