% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtree.r
\name{subtree,phylo-method}
\alias{subtree,phylo-method}
\title{subtree}
\usage{
\S4method{subtree}{phylo}(x, leaves = NULL, ...)
}
\arguments{
\item{x}{a phylo object}

\item{leaves}{leaves of the extracted tree}

\item{...}{additional parameters}
}
\value{
a pylo object
}
\description{
Extract a subtree with specific leaves
}
