% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.r
\name{subtree}
\alias{subtree}
\title{Generic function for subtree}
\usage{
subtree(x, ...)
}
\arguments{
\item{x}{a lineage_tree object}

\item{...}{additional parameters}
}
\description{
Generic function for subtree
}
