% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_kmer.r
\name{summarize_kmer,missing-method}
\alias{summarize_kmer,missing-method}
\title{summarize_kmer}
\usage{
\S4method{summarize_kmer}{missing}(
  x,
  division = 16L,
  k = 2,
  reps = 20L,
  mutation_prob = 0.1,
  alphabets,
  sequence_length = 200L
)
}
\value{
a kmer_summary object
}
\description{
Summarize kmer distributions without input sequences
}
\author{
Wuming Gong (gongx030@umn.edu)
}
