% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_kmer.r
\name{summarize_kmer,phyDat-method}
\alias{summarize_kmer,phyDat-method}
\title{summarize_kmer}
\usage{
\S4method{summarize_kmer}{phyDat}(x, division = 16L, k = 1, reps = 20L, mutation_prob)
}
\description{
Summarize kmer distributions with input sequences
}
