% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_kmer.r
\name{summarize_kmer_core}
\alias{summarize_kmer_core}
\title{summarize_kmer_core}
\usage{
summarize_kmer_core(
  k = 2,
  alphabets,
  mutation_prob,
  division,
  reps,
  sequence_length = 200L,
  outcome_prob,
  n_nodes = 100L,
  dropout_prob = 0
)
}
\arguments{
\item{k}{k-mer (default = 2)}

\item{alphabets}{alphabets used in the tree}

\item{mutation_prob}{mutation probability}

\item{division}{number of cell division}

\item{reps}{number of simulated trees}

\item{sequence_length}{sequence length (e.g. number of targets, default 200L)}

\item{outcome_prob}{outcome probability of each letter}

\item{n_nodes}{number of sampled tip or internval nodes in the simulated tree (default: 100L)}

\item{dropout_prob}{dropout probability (default: 0)}
}
\value{
a kmer_summary object
}
\description{
Summarize kmer distributions (core function)
}
\author{
Wuming Gong (gongx030@umn.edu)
}
