% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WH_train_fit.r
\name{WH_train_fit}
\alias{WH_train_fit}
\title{Train weights for WH, and output distance object}
\usage{
WH_train_fit(x, X)
}
\arguments{
\item{x}{input data in phyDat format}

\item{X}{a list of k number of input data, X[[1]] ... X[[k]]. The ith data have sequence information as phyDat format in X[[i]][[1]], and tree information in X[[i]][[2]] as phylo format.}
}
\value{
a dist object
}
\description{
Train weights for WH using the given data, and fit the distance matrix for a input sequence.
}
\author{
Il-Youp Kwak (ikwak2@cau.ac.kr)
}
