% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.r
\name{as_phylo}
\alias{as_phylo}
\title{Generic function for as_phylo}
\usage{
as_phylo(x, ...)
}
\arguments{
\item{x}{a graph object}

\item{...}{additional parameters}
}
\value{
a phylo object
}
\description{
Generic function for as_phylo
}
