% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_replacment_matrix.r
\name{compute_replacement_matrix,phyDat-method}
\alias{compute_replacement_matrix,phyDat-method}
\title{compute_replacement_matrix}
\usage{
\S4method{compute_replacement_matrix}{phyDat}(
  x,
  n_batch = 10L,
  n_cells_per_batch = 200L,
  k = 2,
  division = 20L,
  dropout = TRUE
)
}
\description{
Simulate replacement matrix with input sequence data
}
