% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.r
\name{dist_weighted_hamming}
\alias{dist_weighted_hamming}
\title{Generic function for dist_weighted_hamming}
\usage{
dist_weighted_hamming(x, wVec, ...)
}
\arguments{
\item{x}{a sequence object}

\item{wVec}{weight vector}

\item{...}{additional parameters}
}
\value{
Calculated distance matrix of input sequences. The result is a 'dist' class object.
}
\description{
Generic function for dist_weighted_hamming
}
