% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_leaves.r
\name{get_leaves,lineage_tree-method}
\alias{get_leaves,lineage_tree-method}
\title{get_leaves}
\usage{
\S4method{get_leaves}{lineage_tree}(x, ...)
}
\arguments{
\item{x}{a lineage_tree object}

\item{...}{additional parameters}
}
\value{
a phyDat object
}
\description{
Get the leaf sequences
}
