% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim.r
\name{prune,lineage_tree-method}
\alias{prune,lineage_tree-method}
\title{prune}
\usage{
\S4method{prune}{lineage_tree}(x, ...)
}
\arguments{
\item{x}{a lineage_tree object}

\item{...}{additional parameters}
}
\value{
a lineage_tree object
}
\description{
Trim a full lineage tree into phylogenetic tree
}
