% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.r
\name{rbind,phyDat-method}
\alias{rbind,phyDat-method}
\title{rbind}
\usage{
\S4method{rbind}{phyDat}(..., deparse.level = 1)
}
\arguments{
\item{...}{a list of phyDat objects}
}
\value{
a phyDat object
}
\description{
Concatenate multiple phyDat objects
}
