% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.r
\name{sample_outcome_prob}
\alias{sample_outcome_prob}
\title{sample_outcome_prob}
\usage{
sample_outcome_prob(config, num_states = 20L, shape = 0.1, scale = 2)
}
\arguments{
\item{shape}{shape parameter in gamma distribution}

\item{scale}{scale parameter in gamma distribution}

\item{alphabets}{alphabets used in the tree}
}
\value{
a probability vector for each alphabet
}
\description{
Sampling outcome probability based on a gamma distribution
}
\author{
Wuming Gong (gongx030@umn.edu)
}
