% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.r
\name{simulate}
\alias{simulate}
\title{simulate}
\usage{
simulate(
  n_samples = 200,
  n_targets = 200,
  mutation_prob = 0.03,
  division = 16L,
  alphabets = NULL,
  outcome_prob = NULL,
  deletion = TRUE
)
}
\description{
Adoped from https://github.com/elifesciences-publications/CRISPR_recorders_sims/blob/master/MATLAB_sims/GESTALT_30hr_1x_simulation.m
}
