% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.r
\name{simulate}
\alias{simulate}
\title{Generic function for simulate}
\usage{
simulate(config, ...)
}
\arguments{
\item{config}{a lineage_tree_config object}

\item{...}{additional parameters}
}
\value{
a lineage_tree object
}
\description{
Generic function for simulate
}
