% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtree.r
\name{subtree,lineage_tree-method}
\alias{subtree,lineage_tree-method}
\title{subtree}
\usage{
\S4method{subtree}{lineage_tree}(x, leaves = NULL, ...)
}
\arguments{
\item{x}{a lineage_tree object}

\item{leaves}{leaves of the extracted tree}

\item{...}{additional parameters}
}
\value{
a lineage_tree object
}
\description{
Extract a subtree with specific leaves
}
