% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_kmer.r
\name{summarize_kmer,phyDat-method}
\alias{summarize_kmer,phyDat-method}
\title{summarize_kmer}
\usage{
\S4method{summarize_kmer}{phyDat}(
  x,
  division = 16L,
  k = 2,
  reps = 20L,
  sequence_length = 200L,
  mutation_prob
)
}
\arguments{
\item{x}{input data as a phyDat object}
}
\value{
a kmer_summary object
}
\description{
Summarize kmer distributions with input sequences
}
\author{
Wuming Gong (gongx030@umn.edu)
}
