% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_kmer.r
\name{summarize_kmer,phyDat-method}
\alias{summarize_kmer,phyDat-method}
\title{summarize_kmer}
\usage{
\S4method{summarize_kmer}{phyDat}(
  x,
  division = 16L,
  k = 2,
  reps = 20L,
  n_samples = 200L,
  n_nodes = 100L,
  n_targets
)
}
\arguments{
\item{x}{input data as a phyDat object}

\item{division}{number of cell division}

\item{k}{k-mer (default = 2)}

\item{reps}{number of simulated trees}

\item{n_samples}{number of samples to simulate}

\item{n_nodes}{number of nodes to sample (including both leaves and internval nodes)}

\item{n_targets}{sequence length. If this argument is missing, the length of the input sequences will be used.}
}
\value{
a kmer_summary object
}
\description{
Summarize kmer distributions with input sequences
}
\author{
Wuming Gong (gongx030@umn.edu)
}
