% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.r
\name{summarize_kmer}
\alias{summarize_kmer}
\title{Generic function for summarize_kmer}
\usage{
summarize_kmer(x, ...)
}
\arguments{
\item{x}{a sequence object}

\item{...}{additional parameters}
}
\description{
Generic function for summarize_kmer
}
