% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_kmer.r
\name{summarize_kmer_core}
\alias{summarize_kmer_core}
\title{summarize_kmer_core}
\usage{
summarize_kmer_core(
  k = 2,
  reps = 20L,
  n_samples = 200L,
  n_nodes = 100L,
  config = NULL
)
}
\arguments{
\item{k}{k-mer (default = 2)}

\item{reps}{number of simulated trees}

\item{n_samples}{number of samples to simulate}

\item{n_nodes}{number of nodes to sample (including both leaves and internval nodes)}

\item{config}{lineage tree configuration (a lineage_tree_config object)}
}
\value{
a kmer_summary object
}
\description{
Summarize kmer distributions (core function)
}
\author{
Wuming Gong (gongx030@umn.edu)
}
